Move compare_dicts, _lists and _values to updater_utils
