+++ tests/simulation/translator/test_surface_meshing_translator.py
@@ -6,6 +6,7 @@ import pytest
 import flow360.component.simulation.units as u
 from flow360.component.simulation.entity_info import GeometryEntityInfo
 from flow360.component.simulation.framework.param_utils import AssetCache
+from flow360.component.simulation.framework.updater_utils import compare_values
 from flow360.component.simulation.meshing_param.edge_params import (
     AngleBasedRefinement,
     HeightBasedRefinement,
@@ -28,7 +29,6 @@ from flow360.component.simulation.unit_system import (
     imperial_unit_system,
 )
 from tests.simulation.conftest import AssetBase
-from tests.utils import compare_values
 
 
 class TempGeometry(AssetBase):
